feat(#69): HttpServletRequest에 userId setAttribute시에 Long으로 타입 통일

+++ server/src/main/java/com/hexacore/tayo/auth/interceptor/AuthenticationInterceptor.java
@@ -39,7 +39,7 @@ public class AuthenticationInterceptor implements HandlerInterceptor {
 
         // 엑세스 토큰의 만료여부를 확인
         if (claims.getExpiration().after(new Date())) { // 만료되지 않은 토큰인 경우
-            request.setAttribute("userId", claims.get("userId"));
+            request.setAttribute("userId", Long.valueOf((Integer) claims.get("userId")));
             request.setAttribute("userName", claims.get("userName"));
             return true;
         } else { // 만료된 토큰인 경우
